Added ServiceLocator and unit/instrum tests with mock API
For this simple app, I preferred to use a simple ServiceLocator rather
than a complex Dagger 2 structure with generated code and annotations.

The `ServiceLocator` interface defines the services dependencies used
throughout the app. It has a singleton instance which can be referred to
as the default value for constructor dependency injection, like done for
the `ListPresenter`.

 `AppServiceLocator` is a singleton class implementing `ServiceLocator`.
It is assigned as the `ServiceLocator` instance by `CustomApplication`,
defined as this app's `Application` in the manifest.

In instrumentation tests, we have a `TestApplication` instead, which
assigns a `TestServiceLocator`. It uses an implementation of
`TMDBService` with an underlying `MockWebServer`, which accepts mock
responses to be enqueued (and returns them to the requests made to the
service, in the same order in which they were enqueued).

In order to reuse the `TestServiceWrapper` and the local mocked JSONs
between instrumentation and unit tests, they were added to a
`sharedTest` directory, which is made available to `test` and
`androidTest` in the gradle config. Curiously, the JSON resource file
wasn't found when running instrumentation tests, even though it was in
the sharedTest directory and even though it worked fine for unit tests.
The solution was to replace `getResource` by `getResourceAsStream`.