Instantiates ListViewModel via ViewModelProviders
This is a step towards including landscape support. We have a ViewModel
because we want it to retain its state after rotation, but for that we
also need the ViewModel object to remain the same after rotation, which
is what ViewModelProviders is doing for us. Using `by lazy` to avoid the
exception "Can't create ViewModelProvider for detached fragment".

We lost the ability to inject the service into the ViewModel, because
otherwise ViewModelProviders wouldn't be able to call our custom
constructor, but this service parameter wasn't being used anyway (we are
defining our own ServiceLocator when we need mock dependencies).

We also got rid of the ListContract now, because ViewModelProviders
requires an actual `ViewModel` which is not an interface but an abstract
class.